requirements.txt
openai-whisper
googletrans==4.0.0rc1
gTTS
pydub
